Migrate query string examples to new structure
