+++ examples/query_string/README.md
@@ -1,0 +1,35 @@
+# Query String Examples
+
+A collection of crates that provide examples of working with the request query string and the
+Gotham web framework.
+
+## Help
+
+Our welcoming community is available to help:
+
+* [The Gotham web framework website](https://gotham.rs)
+* [Gotham web framework API documentation](https://docs.rs/gotham/)
+* [Gitter chatroom](https://gitter.im/gotham-rs/gotham)
+* [Twitter](https://twitter.com/gotham_rs)
+
+## Ordering
+
+We recommend reviewing our request query string examples in the order shown below:
+
+1. [Introduction](introduction) - Introduces having the Gotham web framework Router extract request
+                                  query string name/value pairs, in a type safe way, for use in
+                                  your application.
+
+## License
+
+Licensed under your option of:
+
+* [MIT License](../LICENSE-MIT)
+* [Apache License, Version 2.0](../LICENSE-APACHE)
+
+## Community
+
+The following policies guide participation in our project and our community:
+
+* [Conduct](../../CONDUCT.md)
+* [Contributing](../../CONTRIBUTING.md)
